Swap efficiencies and fitting chapters.
I think fitting first makes more sense, because one of the efficiencies
is due to a procedure applied for the fit. It’s confusing to talk
about something that’s later in the note, with sufficient detail for the reader to understand enough, but without repeating what will be said
later.